+++ client/control/components/Header.js
@@ -1,8 +1,8 @@
-import React, { Component, PropTypes as pt } from 'react';
+import React, { PropTypes as pt } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router';
 import Breadcrumbs from 'react-breadcrumbs';
-import QueryAPIToken from './data/QueryAPIToken.js';
+import APITokenExpander from './APITokenExpander.js';
 
 function Header({
   pageType: { ctaButtons },
@@ -62,46 +62,3 @@ export default connect(
     token: state.auth.token,
   }),
 )(Header);
-
-
-class APITokenExpander extends Component {
-  static propTypes = {
-    token: pt.shape({ key: pt.string }),
-  };
-
-  constructor(props) {
-    super(props);
-    this.state = { expanded: false };
-    this.toggleExpand = ::this.toggleExpand;
-  }
-
-  toggleExpand() {
-    this.setState(state => ({ ...state, expanded: !state.expanded }));
-  }
-
-  render() {
-    const { token } = this.props;
-    const { expanded } = this.state;
-
-    const expandButton = (
-      <button className="button-icon" onClick={this.toggleExpand}>
-        <i className="fa fa-gear fa-lg" />
-      </button>
-    );
-
-    if (expanded) {
-      return (
-        <span>
-          <QueryAPIToken />
-          {token && `API Key: ${token.key}`}
-          {expandButton}
-        </span>
-      );
-    }
-    return (
-      <span>
-        {expandButton}
-      </span>
-    );
-  }
-}
