+++ client/control/sass/control.scss
@@ -276,5 +276,4 @@
   border: none;
   cursor: pointer;
   padding: 0;
-  transform: translate(0, -1px);
 }
